Added example code
Added a "full cycle" example using the Strava PHP API.

This example presents a page with a "Connect With Strava" button
Clicking that button requests authorization from Strava to access
as Strava user's data.

When authorization is given, the user is returned to the requesting
app and is presented with data from Strava.